implementação do auth

+++ app/Http/Controllers/ContatosController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\Contato;
 use Session;
@@ -16,21 +17,19 @@ class ContatosController extends Controller
     public function index()
     {
         $contatos = Contato::paginate(5);
-        return view(('contato.index'),array('contatos' => $contatos, 'busca' => null));
+        return view('contato.index',array('contatos' => $contatos,'busca'=>null));
     }
 
     /**
-     * Show the form for creating a new resource.
+     * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-
     public function buscar(Request $request) {
-        $contatos = Contato::where('nome', 'LIKE', '%'.$request->input('busca')
-        .'%')->orwhere('email', 'LIKE', '%'.$request->input('busca').'%')->get();
-        return view('contato.index', array('contatos' => $contatos, 'busca' => $request->input('busca')));
-        
+        $contatos = Contato::where('nome','LIKE','%'.$request->input('busca').'%')->orwhere('email','LIKE','%'.$request->input('busca').'%')->paginate(5);
+        return view('contato.index',array('contatos' => $contatos,'busca'=>$request->input('busca')));
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -39,7 +38,13 @@ class ContatosController extends Controller
      */
     public function create()
     {
-        return view('contato.create');
+        if (Auth::check()) {
+            return view('contato.create');
+        }
+        else {
+            return redirect('login');
+        }
+
     }
 
     /**
@@ -50,26 +55,31 @@ class ContatosController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[  
-            'nome' => 'required|min:3',
-            'email' => 'required|e-mail',
-            'telefone' => 'required',
-            'cidade' => 'required',
-            'estado' => 'required',
-        ]);
-        $contato = new Contato();
-        $contato->nome = $request->input('nome');
-        $contato->email = $request->input('email');
-        $contato->telefone = $request->input('telefone');
-        $contato->cidade = $request->input('cidade');
-        $contato->estado = $request->input('estado');
-        if($contato->save()) {
-            if($request->hasFile('foto')){
-                $imagem = $request->file('foto');
-                $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
-                $request->file('foto')->move(public_path('.\img\contatos'),'nomearquivo');
+        if (Auth::check()) {
+            $this->validate($request,[
+                'nome' => 'required|min:3',
+                'email' => 'required|e-mail',
+                'telefone' => 'required',
+                'cidade' => 'required',
+                'estado' => 'required',
+            ]);
+            $contato = new Contato();
+            $contato->nome = $request->input('nome');
+            $contato->email = $request->input('email');
+            $contato->telefone = $request->input('telefone');
+            $contato->cidade = $request->input('cidade');
+            $contato->estado = $request->input('estado');
+            if($contato->save()) {
+                if($request->hasFile('foto')){
+                    $imagem = $request->file('foto');
+                    $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
+                    //dd($imagem, $nomearquivo,$contato->id);
+                    $request->file('foto')->move(public_path('.\img\contatos'),$nomearquivo);
+                }
+                return redirect('contatos');
             }
-            return redirect('contatos');
+        } else {
+            return redirect('login');
         }
     }
 
@@ -82,7 +92,7 @@ class ContatosController extends Controller
     public function show($id)
     {
         $contato = Contato::find($id);
-        return view(('contato.show'), array('contato' => $contato));
+        return view('contato.show',array('contato' => $contato));
     }
 
     /**
@@ -93,8 +103,12 @@ class ContatosController extends Controller
      */
     public function edit($id)
     {
-        $contato = Contato::find($id);
-        return view('contato.edit', array('contato' => $contato));
+        if (Auth::check()) {
+            $contato = Contato::find($id);
+            return view('contato.edit',array('contato' => $contato));
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -106,44 +120,53 @@ class ContatosController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request,[  
-            'nome' => 'required|min:3',
-            'email' => 'required|e-mail',
-            'telefone' => 'required',
-            'cidade' => 'required',
-            'estado' => 'required',
-        ]);
-        $contato = Contato::find($id);
-        if($request->hasFile('foto')){
-            $imagem = $request->file('foto');
-            $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
-            $request->file('foto')->move(public_path('.\img\contatos'),$nomearquivo);
-        }
-        $contato->nome = $request->input('nome');
-        $contato->email = $request->input('email');
-        $contato->telefone = $request->input('telefone');
-        $contato->cidade = $request->input('cidade');
-        $contato->estado = $request->input('estado');
-        if($contato->save()) {
-            Session::flash('mensagem', 'Contato alterado com sucesso');
-            return view('contato.show', array('contato' => $contato));
+        if (Auth::check()) {
+            $this->validate($request,[
+                'nome' => 'required|min:3',
+                'email' => 'required|e-mail|min:3',
+                'telefone' => 'required',
+                'cidade' => 'required',
+                'estado' => 'required',
+            ]);
+            $contato = Contato::find($id);
+            if($request->hasFile('foto')){
+                $imagem = $request->file('foto');
+                $nomearquivo = md5($contato->id).".".$imagem->getClientOriginalExtension();
+                $request->file('foto')->move(public_path('.\img\contatos'),$nomearquivo);
+            }
+            $contato->nome = $request->input('nome');
+            $contato->email = $request->input('email');
+            $contato->telefone = $request->input('telefone');
+            $contato->cidade = $request->input('cidade');
+            $contato->estado = $request->input('estado');
+            if($contato->save()) {
+                Session::flash('mensagem','Contato alterado com sucesso');
+                return redirect()->back();
+            }
+        } else {
+            return redirect('login');
         }
     }
 
     /**
      * Remove the specified resource from storage.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy(Request $request, $id)
     {
-        $contato = Contato::find($id);
-        if (isset($request->foto)) {
+        if (Auth::check()) {
+            $contato = Contato::find($id);
+            if (isset($request->foto)) {
             unlink($request->foto);
+            }
+            $contato->delete();
+            Session::flash('mensagem','Contato Excluído com Sucesso Foto:');
+            return redirect(url('contatos/'));
+        } else {
+            return redirect('login');
         }
-        $contato->delete();
-        Session::flash('mensagem', 'Contato Excluído com Sucesso!');
-        return redirect(url('contatos/'));
     }
 }
