+++ app/Http/Controllers/EmprestimoController.php
@@ -2,13 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Models\Emprestimo;
 use App\Models\Contato;
 use App\Models\Livro;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Session;
 
-class EmprestimoController extends Controller
+class EmprestimosController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -19,7 +21,7 @@ class EmprestimoController extends Controller
     {
         $emprestimos = Emprestimo::simplepaginate(5);
         return view('emprestimo.index',array('emprestimos' => $emprestimos,'busca'=>null));
-
+        return redirect('login');
     }
 
     /**
@@ -28,7 +30,14 @@ class EmprestimoController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function buscar(Request $request) {
-        $emprestimos = Emprestimo::where('contato_id','=',$request->input('busca'))->orwhere('livro_id','=',$request->input('busca'))->orwhere('obs','LIKE','%'.$request->input('busca').'%')->simplepaginate(5);
+        $emprestimos = Emprestimo::join('contatos','contatos.id','=','emprestimos.contato_id')
+                    ->join('livros','livros.id','=','emprestimos.livro_id')
+                    ->select('emprestimos.*','contatos.nome','livros.titulo')
+                    ->where('contato_id','=',$request->input('busca'))
+                    ->orwhere('livro_id','=',$request->input('busca'))
+                    ->orwhere('obs','LIKE','%'.$request->input('busca').'%')->orwhere('contatos.nome','LIKE','%'.$request->input('busca').'%')
+                    ->orwhere('livros.titulo','LIKE','%'.$request->input('busca').'%')
+                    ->simplepaginate(5);
         return view('emprestimo.index',array('emprestimos' => $emprestimos,'busca'=>$request->input('busca')));
     }
 
@@ -40,9 +49,13 @@ class EmprestimoController extends Controller
      */
     public function create()
     {
-        $contatos = Contato::all();
-        $livros = Livro::all();
-        return view('emprestimo.create',['contatos'=>$contatos,'livros'=>$livros]);
+        if (Auth::check()) {
+            $contatos = Contato::all();
+            $livros = Livro::all();
+            return view('emprestimo.create',['contatos'=>$contatos,'livros'=>$livros]);
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -53,20 +66,24 @@ class EmprestimoController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[
-            'contato_id' => 'required',
-            'livro_id' => 'required',
-            'datahora' => 'required'
-        ]);
-        $emprestimo = new Emprestimo();
-        $emprestimo->contato_id = $request->input('contato_id');
-        $emprestimo->livro_id = $request->input('livro_id');
-        $emprestimo->datahora = \Carbon\Carbon::createFromFormat('d/m/Y H:i:s', $request->input('datahora'));
-        $emprestimo->obs = $request->input('obs');
-        $emprestimo->datadevolucao = null;
+        if (Auth::check()) {
+            $this->validate($request,[
+                'contato_id' => 'required',
+                'livro_id' => 'required',
+                'datahora' => 'required'
+            ]);
+            $emprestimo = new Emprestimo();
+            $emprestimo->contato_id = $request->input('contato_id');
+            $emprestimo->livro_id = $request->input('livro_id');
+            $emprestimo->datahora = \Carbon\Carbon::createFromFormat('d/m/Y H:i:s', $request->input('datahora'));
+            $emprestimo->obs = $request->input('obs');
+            $emprestimo->datadevolucao = null;
 
-        if($emprestimo->save()) {
-            return redirect('emprestimos');
+            if($emprestimo->save()) {
+                return redirect('emprestimos');
+            }
+        } else {
+            return redirect('login');
         }
     }
 
@@ -83,6 +100,7 @@ class EmprestimoController extends Controller
         return view('emprestimo.show',array('emprestimo' => $emprestimo,'busca'=>null));
     }
 
+
     /**
      * Show the form for editing the specified resource.
      *
@@ -92,6 +110,29 @@ class EmprestimoController extends Controller
     public function edit(Emprestimo $emprestimo)
     {
         //
+    }
+
+/**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int $id
+    *  @return \Illuminate\Http\Response
+     */
+    public function devolver(Request $request, $id)
+    {
+        if (Auth::check()) {
+            $emprestimo = Emprestimo::find($id);
+            $emprestimo->datadevolucao = \Carbon\Carbon::now();
+            $emprestimo->save();
+
+            if($emprestimo->save()) {
+                Session::flash('mensagem','Empréstimo Devolvido');
+                return redirect()->back();
+            }
+        } else {
+            return redirect('login');
+        }
     }
 
     /**
@@ -108,16 +149,20 @@ class EmprestimoController extends Controller
 
     /**
      * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Emprestimo  $emprestimo
+     * @param \Illuminate\Http\Request $request
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Emprestimo $emprestimo, $id)
+    public function destroy(Request $request, $id)
     {
-        $emprestimo = Emprestimo::find($id);
+        if (Auth::check()) {
+            $emprestimo = Emprestimo::find($id);
 
-        $emprestimo->delete();
-        Session::flash('mensagem','Empréstimo Excluído com Sucesso');
-        return redirect(url('emprestimos/'));
+            $emprestimo->delete();
+            Session::flash('mensagem','Empréstimo Excluído com Sucesso');
+            return redirect(url('emprestimos/'));
+        } else {
+            return redirect('login');
+        }
     }
 }
